Added custom fonts + temperature graph.

+++ app/src/main/java/com/example/androiddevchallenge/model/HourlyForecast.kt
@@ -1,3 +1,3 @@
 package com.example.androiddevchallenge.model
 
-data class HourlyForecast(val temperature: Double, val state: WeatherState)
+data class HourlyForecast(val temperature: Float, val state: WeatherState)
